pyproject.toml
[project]
name = "heptapal-agent-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.6.1",
    "pyyaml>=6.0",
    "mysql-connector-python>=8.2.0",
    "sqlalchemy>=2.0.0",
]
